requirements.txt
python-telegram-bot==22.7
google-ads==29.2.0
anthropic>=0.34.0
APScheduler==3.10.4
python-dotenv==1.0.1
